Moves desktop list item into desktop table folder
