+++ components/DesktopInventoryItemList/DesktopInventoryItemListItem/index.tsx
